requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pandas>=1.5.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
numpy>=1.24.0
boto3>=1.26.0
PyJWT==2.8.0 